Add league and team schedule windows
